hw2-handout/src/requirements.txt
numpy
pandas
tqdm
matplotlib
seaborn
torch
scipy
scikit-learn
liblinear-official
jupyter